巴士saas-bug修复

+++ src/main/java/com/zhuanche/vo/busManage/BusBizDriverViolatorsVO.java
@@ -42,7 +42,7 @@ public class BusBizDriverViolatorsVO {
 
     private String punishReason;
 
-    private Integer punishDuration;
+    private Double punishDuration;
 
     private Date punishStartTime;
